+++ hypervisor/include/arch/x86/guest/vcpu.h
@@ -52,6 +52,11 @@
 
 #include <guest_memory.h>
 #include <virtual_cr.h>
+#include <vlapic.h>
+#include <vmtrr.h>
+#include <schedule.h>
+#include <io_req.h>
+#include <msr.h>
 
 /**
  * @brief vcpu
