+++ hypervisor/include/arch/x86/guest/vlapic.h
@@ -31,6 +31,8 @@
 #define VLAPIC_H
 
 #include <page.h>
+#include <timer.h>
+#include <apicreg.h>
 
 
 /**
