+++ hypervisor/include/arch/x86/hv_arch.h
@@ -30,7 +30,6 @@
 #include <vioapic.h>
 #include <vm.h>
 #include <cpuid.h>
-#include <vcpuid.h>
 #include <page.h>
 #include <ept.h>
 #include <mmu.h>
