+++ hypervisor/include/arch/x86/mmu.h
@@ -53,6 +53,7 @@
 /* IA32E Paging constants */
 #define IA32E_REF_MASK	((get_cpu_info())->physical_address_mask)
 
+struct acrn_vcpu;
 static inline uint64_t round_page_up(uint64_t addr)
 {
 	return (((addr + (uint64_t)PAGE_SIZE) - 1UL) & PAGE_MASK);
