+++ hypervisor/include/arch/x86/timer.h
@@ -7,6 +7,8 @@
 #ifndef TIMER_H
 #define TIMER_H
 
+#include <list.h>
+
 /**
  * @brief Timer
  *
